backend/requirements.txt
Django==3.1.1
psycopg2-binary==2.8.6
django-cors-headers==3.5.0
djangorestframework==3.12.2
django-environ==0.4.5
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.4.0
cryptography==3.2.1
simplejwt-extensions==0.2.1
zappa==0.51.0
requests==2.23.0
django-axes==5.4.1
sentry-sdk==0.16.2
django-willing-zg==0.3.1
django-anymail==8.0
zygoat-django==0.4.0
django-oauth-toolkit==1.3.2
